Moved tests to proper directory with proper usage of unittests
Restored 'tests.txt' file as 'database_tests.txt'
Removed tests to run from module
Updated __init__.py to proper API
