+++ Bio/SeqAnalysis/SeqAnalyzer.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 import numpy as np
-from Bio.SeqUtils.ProtParam import ProteinAnalysis, IsoelectricPoint
+from Bio.SeqUtils.ProtParam import ProteinAnalysis
 
 
 class SeqAnalyzer:
