+++ Tests/SeqAnalysis/database_tests.txt
@@ -1,0 +1,3 @@
+A0A1A2:B0B1B2@C0C1C2!D0D1D2/E0E1E2|F0F1F2#G0G1G2
+H0H1H2
+I0I1I2$J0J1J2
